Minimizer type changed from uint64_t to uint32_t

+++ clustering/cluster.h
@@ -10,7 +10,7 @@
 #define CLUSTER_H
 
 typedef size_t read_id_t;
-typedef uint64_t minimizer_t;
+typedef uint32_t minimizer_t;
 
 extern size_t node_count;
 extern size_t read_count;
